+++ tests/test-files/interface/nok.ts
@@ -1,0 +1,13 @@
+interface Moo {
+    x: number
+    Y: string
+    invalid_property: boolean;
+}
+
+const moo: Moo = {
+    x: 5,
+    Y: '6',
+    invalid_property: false,
+}
+
+void moo