+++ tests/test-files/interface/ok.ts
@@ -1,0 +1,11 @@
+interface IMoo {
+    x: number;
+    Y: string;
+}
+
+const moo: IMoo = {
+    Y: '6',
+    x: 5,
+};
+
+void moo;