+++ tests/test-files/nok/interface.ts
@@ -1,5 +1,0 @@
-interface Moo {
-    x: number
-    Y: string
-    invalid_property: boolean;
-}