+++ tests/test-files/ok/interface.ts
@@ -1,4 +1,0 @@
-interface IMoo {
-    x: number;
-    Y: string;
-}