+++ tests/test.js
@@ -1,7 +1,9 @@
 const fs = require('fs');
+const path = require('path');
+
 const eslint = require('eslint');
 
-function getErrors(fileToTest, project = 'tests/configs/test.tsconfig.json') {
+function getErrors(fileToTest, project = path.join('tests', 'configs', 'test.tsconfig.json')) {
     const CLIEngine = eslint.CLIEngine;
 
     const cli = new CLIEngine({
@@ -16,88 +18,23 @@ function getErrors(fileToTest, project = 'tests/configs/test.tsconfig.json') {
 describe('Self-lint', () => {
 
     it('must not have any errors in Index.js', () => {
-        expect(getErrors('index.js', 'tests/configs/index.tsconfig.json').results[0].messages).toEqual([]);
+        expect(getErrors('index.js', path.join('tests', 'configs', 'index.tsconfig.json')).results[0].messages).toEqual([]);
     });
 });
 
-describe('Validate ESLint configs on valid files', () => {
+describe('Validate ESLint configs on files', () => {
 
-    const path = 'tests/test-files/ok/';
-    const files = fs.readdirSync(path);
+    const filesPath = path.join('tests', 'test-files');
+    const contents = fs.readdirSync(filesPath, {withFileTypes: true});
+    const directories = contents.filter((content) => content.isDirectory()).map((dir) => dir.name);
 
-    files.forEach((file) => {
-        it(file, () => {
-            expect(getErrors(path + file).results[0].messages).toEqual([]);
-        });
+    it.each(directories)('directory: %s', (directory) => {
+        const errorFilePath = path.join(filesPath, directory, 'errors.json');
+        const notOkFile = path.join(filesPath, directory, 'nok.ts');
+        const okFile = path.join(filesPath, directory, 'ok.ts');
+
+        expect(getErrors(okFile).results[0].messages).toEqual([]);
+        expect(getErrors(notOkFile).results[0].messages).toEqual(JSON.parse(fs.readFileSync(errorFilePath).toString()));
     });
-});
 
-describe('Validate ESLint configs on invalid files', () => {
-
-    const path = 'tests/test-files/nok/';
-
-    it('interfaces', () => {
-        expect(getErrors(path + 'interface.ts').results[0].messages).toEqual(
-            [
-                {
-                    "column": 11,
-                    "endColumn": 14,
-                    "endLine": 1,
-                    "line": 1,
-                    "message": "Interface name `Moo` must have one of the following prefixes: I",
-                    "messageId": "missingAffix",
-                    "nodeType": "Identifier",
-                    "ruleId": "@typescript-eslint/naming-convention",
-                    "severity": 2
-                },
-                {
-                    "column": 14,
-                    "endColumn": 14,
-                    "endLine": 2,
-                    "fix": {
-                        "range": [
-                            29,
-                            29
-                        ],
-                        "text": ";"
-                    },
-                    "line": 2,
-                    "message": "Expected a semicolon.",
-                    "messageId": "expectedSemi",
-                    "nodeType": "Identifier",
-                    "ruleId": "@typescript-eslint/member-delimiter-style",
-                    "severity": 2
-                },
-                {
-                    "column": 14,
-                    "endColumn": 14,
-                    "endLine": 3,
-                    "fix": {
-                        "range": [
-                            43,
-                            43
-                        ],
-                        "text": ";"
-                    },
-                    "line": 3,
-                    "message": "Expected a semicolon.",
-                    "messageId": "expectedSemi",
-                    "nodeType": "Identifier",
-                    "ruleId": "@typescript-eslint/member-delimiter-style",
-                    "severity": 2
-                },
-                {
-                    "column": 5,
-                    "endColumn": 21,
-                    "endLine": 4,
-                    "line": 4,
-                    "message": "Identifier 'invalid_property' is not in camel case.",
-                    "messageId": "notCamelCase",
-                    "nodeType": "Identifier",
-                    "ruleId": "camelcase",
-                    "severity": 2
-                }
-            ]
-        );
-    });
 });